GPII-1783:  Rearranged for easier reuse per PR feedback.
