--- tests/js/includes.js
@@ -1,15 +1,0 @@
-"use strict";
-var fluid = require("infusion");
-var gpii  = fluid.registerNamespace("gpii");
-
-fluid.loadTestingSupport();
-
-var kettle = require("kettle");
-kettle.loadTestingSupport();
-
-require("gpii-express");
-gpii.express.loadTestingSupport();
-
-require("./harness");
-require("./environment");
-require("./lib/helpers");
