+++ tests/js/launch-test-harness.js
@@ -3,11 +3,18 @@
 var fluid = require("infusion");
 var gpii  = fluid.registerNamespace("gpii");
 
-require("./harness");
+require("./lib/harness");
+require("./pouch-config.js");
 
 fluid.setLogging(true);
 
 gpii.test.pouch.harness({
-    port:       6789,
-    baseUrl:    "http://localhost:6789/"
+    port: 6789,
+    databases: gpii.tests.pouch.config.databases,
+    distributeOptions: [
+        {
+            source: "{that}.options.databases",
+            target: "{that gpii.pouch}.options.databases"
+        }
+    ]
 });
