+++ tests/js/lib/harness.js
@@ -4,18 +4,10 @@
 var fluid = require("infusion");
 
 require("gpii-express");
-require("../../");
-
+fluid.require("%gpii-pouchdb");
 
 fluid.defaults("gpii.test.pouch.harness", {
     gradeNames: ["fluid.component"],
-    databases: {
-        sample:  { data: [ "%gpii-pouchdb/tests/data/data.json", "%gpii-pouchdb/tests/data/supplemental.json"] },
-        _users:  { data: "%gpii-pouchdb/tests/data/users.json"},
-        // A ~100k data set to confirm that the async data loads do not take too long.
-        massive: { data: "%gpii-pouchdb/tests/data/massive.json"},
-        nodata:  {}
-    },
     events: {
         expressStarted: null,
         pouchStarted:   null,
@@ -31,7 +23,6 @@ fluid.defaults("gpii.test.pouch.harness", {
             type: "gpii.express",
             options: {
                 "port" : "{harness}.options.port",
-                baseUrl: "{harness}.options.baseUrl",
                 listeners: {
                     onStarted: "{harness}.events.expressStarted.fire"
                 },
@@ -40,7 +31,6 @@ fluid.defaults("gpii.test.pouch.harness", {
                         type: "gpii.pouch",
                         options: {
                             path: "/",
-                            databases: "{harness}.options.databases",
                             listeners: {
                                 onStarted: "{harness}.events.pouchStarted.fire"
                             }
